requirements.txt
colorama
httpx
tls_client
websocket_client
